TST: renamed file with shared test functions
