+++ bluesky_queueserver/manager/tests/test_start_re_manager_cli.py
@@ -1,9 +1,9 @@
 import pytest
 
 from ..comms import zmq_single_request
-from ._common import re_manager_cmd  # noqa: F401
+from .common import re_manager_cmd  # noqa: F401
 
-from ._common import (
+from .common import (
     wait_for_condition,
     condition_environment_created,
     condition_environment_closed,
