+++ bluesky_queueserver/server/tests/test_http_server_func_scope.py
@@ -4,7 +4,7 @@ import pytest
 
 from ._common import plan_list_sample, create_excel_file_from_plan_list
 
-from bluesky_queueserver.manager.tests._common import (  # noqa F401
+from bluesky_queueserver.manager.tests.common import (  # noqa F401
     re_manager,
     re_manager_pc_copy,
     re_manager_cmd,
